Made pages prettier, added startup animation, loving, crys for help.
The girlfriend object now pours out it's love, unless it's
separated in which case it pours out pleas for help.

The Big Bad bides his time, swoops in and kidnaps the girlfriend,
keeping her just off the top of the screen until the top of the
level.

The pages have been made prettier, and a php page (not included in
the archive) added which makes the whole thing appear from
within my website rather than just a page on it's own.

"Joust Adventure" is the name, and has a logo made of CSS and
IMPACT. Which is rubbish, but still better.