Level Completion Code, links for next level etc.
We fire off the level complete stuff when we reach the top
and get close to the rescue. That stops the game apart from
the pretty particle text things saying "Complete" and then
when you click we go to level2.php

Note we're in php all the time now, though I might go back
to a HTML version later, at last add one.