Added completion indicator and fixed a bug.
The bird you're riding now always faces the same was as you,
or more exactly, the big bad's bird never gets confused that
way which it used to quite a lot.

Also added a completion indicator. Think that'll encourage
folks to get it up to 100% and so see the next level begging
plea.